requirements.txt
streamlit>=1.40
langchain>=0.3
langchain-community>=0.3
langchain-huggingface>=0.1
langchain-text-splitters>=0.3
supabase>=2.0
pypdf>=5.0
python-dotenv>=1.0
langchain-openai>=0.1